requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
openai==1.10.0
tiktoken==0.5.2
